+++ Source/Reports/Children/InitialisationPartition.cpp
@@ -61,7 +61,7 @@ void InitialisationPartition::DoExecute() {
     cache_ << " " << i;
   cache_ << "\n";
 
-  vector<unsigned> age_freq;
+  vector<float> age_freq;
   for (unsigned row = 0; row < model_->get_height(); ++row) {
     for (unsigned col = 0; col < model_->get_width(); ++col) {
       WorldCell* cell = world_->get_base_square(row, col);
